+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [unreleased]
+### Changed
+- Disabled Kukai Embed silent signing
+- Updated Kukai Embed to `0.8.9`
+
+
 ## [2.0.5] - 2023-08-12
 ### Changed
 - Unity readme